requirements.txt
flask[async]
flask-socketio
pyserial
simple-websocket
threaded